episode_id long 로직 처리, 영상 재생 테스트 위한 sql문 추가

+++ src/main/java/com/example/vickey/api/EpisodeController.java
@@ -86,7 +86,7 @@ public class EpisodeController {
     }
 
     @GetMapping({"/contentInfo"})
-    public Episode contentInfoEpisodes(@RequestParam Integer contentInfoQuery) {
+    public Episode contentInfoEpisodes(@RequestParam Long contentInfoQuery) {
         return episodeService.contentInfoEpisodes(contentInfoQuery);
     }
 }
